(fix) Make isASTNode function work with prettier 2.1.x
This change alters the code such that we are independent of property-changes of both Svelte and Prettier by adding a `__isRoot`-property to the parser output root.
Fixes #123
