+++ .gitignore
@@ -1,4 +1,5 @@
 node_modules
 .vscode
+.history
 plugin.js
 plugin.js.map
