+++ package.json
@@ -29,9 +29,9 @@
     "homepage": "https://github.com/sveltejs/prettier-plugin-svelte#readme",
     "devDependencies": {
         "@types/node": "^10.12.18",
-        "@types/prettier": "^2.0.0",
+        "@types/prettier": "^2.0.2",
         "ava": "1.2.0",
-        "prettier": "^2.0.0",
+        "prettier": "^2.1.0",
         "rollup": "1.1.2",
         "rollup-plugin-commonjs": "9.2.0",
         "rollup-plugin-node-resolve": "4.0.0",
